+++ src/modules/Home.service.js
@@ -18,8 +18,6 @@ const pagination = document.querySelector('.Pagination');
 
 const HomePage = async () => {
   const movieListData = await getMovieHandler();
-  // await movieData.init()
-  // const movieListData = movieData.data
   if (movieListData.length < 0) {
     // ? could use spinner component here when fetching data
     movieLists.innerHTML = 'loading';
@@ -49,10 +47,6 @@ const HomePage = async () => {
   reservations();
   // renderCommentPopup()
   commentListener();
-
-  // homeCommentBtn.forEach((btn) => {
-  //   btn.addEventListener("click", renderCommentPopup)
-  // })
 
   // - TODO: end
   const previousButton = document.querySelector('.previousPage');
